@@ -10,22 +10,28 @@
     "build": "node build/build.js",
     "build:report": "npm_config_report=true node build/build.js",
     "lint": "eslint --ext .js,.vue src",
+    "lintfix": "eslint --fix  --ext  .js,.vue src",
     "test": "npm run lint"
   },
   "dependencies": {
+    "animate.css": "^3.7.0",
     "axios": "0.18.0",
+    "earcut": "^2.1.5",
     "echarts": "^4.1.0",
-    "element-ui": "2.3.4",
+    "element-ui": "^2.6.3",
+    "mockjs": "^1.0.1-beta3",
     "js-cookie": "2.2.0",
+    "moment": "^2.24.0",
     "node-sass": "^4.11.0",
     "normalize.css": "7.0.0",
     "nprogress": "0.2.0",
     "sockjs-client": "^1.3.0",
     "stompjs": "^2.3.3",
     "three": "^0.101.1",
-    "vue": "2.5.10",
+    "vue": "2.5.16",
     "vue-baidu-map": "^0.21.16",
     "vue-router": "3.0.1",
+    "vue-ueditor-wrap": "^2.4.1",
     "vuex": "3.0.1",
     "websocket": "^1.0.28"
   },
@@ -43,10 +49,16 @@
     "chalk": "2.3.0",
     "copy-webpack-plugin": "4.2.3",
     "css-loader": "0.28.7",
-    "eslint": "4.13.1",
+    "eslint": "^4.19.1",
+    "eslint-config-standard": "^12.0.0",
     "eslint-friendly-formatter": "3.0.0",
     "eslint-loader": "1.9.0",
     "eslint-plugin-html": "4.0.1",
+    "eslint-plugin-import": "^2.16.0",
+    "eslint-plugin-node": "^8.0.1",
+    "eslint-plugin-promise": "^4.1.1",
+    "eslint-plugin-standard": "^4.0.0",
+    "eslint-plugin-vue": "^5.2.2",
     "eventsource-polyfill": "0.9.6",
     "extract-text-webpack-plugin": "3.0.2",
     "file-loader": "1.1.5",
@@ -67,7 +79,7 @@
     "url-loader": "0.6.2",
     "vue-loader": "13.5.0",
     "vue-style-loader": "3.0.3",
-    "vue-template-compiler": "2.5.10",
+    "vue-template-compiler": "2.5.16",
     "webpack": "3.10.0",
     "webpack-bundle-analyzer": "2.9.1",
     "webpack-dev-server": "2.9.7",
