@@ -5,21 +5,21 @@
 </template>
 
 <script>
-  export default {
-    name: 'app',
-    methods: {
-      handler: function () {
+export default {
+  name: 'app',
+  methods: {
+    handler: function () {
 
-      }
-    },
-    mounted () {
-      // this.websocket.initWebSocket()
-      document.addEventListener('mousewheel', this.handler, {passive: true});
-    },
-    beforeDestroy () {
-      // this.websocket.close()
     }
+  },
+  mounted () {
+    // this.websocket.initWebSocket()
+    document.addEventListener('mousewheel', this.handler, { passive: true })
+  },
+  beforeDestroy () {
+    // this.websocket.close()
   }
+}
 </script>
 
 <style lang="scss">
@@ -27,6 +27,7 @@
   @import './styles/index.scss'; // 全局自定义的css样式
 
   body {
-    background: url('../src/images/paper.gif')
+    /*background: url('../src/images/paper.gif')*/
+    background: #F5F5F5;
   }
 </style>
