@@ -3,537 +3,427 @@
  */
 import * as THREE from 'three'
 
-export default function TrackballControls( object, domElement ) {
+export default function TrackballControls (object, domElement) {
+  THREE.EventDispatcher.call(this)
 
-	THREE.EventDispatcher.call( this );
+  var _this = this
+  var STATE = { NONE: -1, ROTATE: 0, ZOOM: 1, PAN: 2, TOUCH_ROTATE: 3, TOUCH_ZOOM: 4, TOUCH_PAN: 5 }
 
-	var _this = this;
-	var STATE = { NONE: -1, ROTATE: 0, ZOOM: 1, PAN: 2, TOUCH_ROTATE: 3, TOUCH_ZOOM: 4, TOUCH_PAN: 5 };
+  this.object = object
+  this.domElement = (domElement !== undefined) ? domElement : document
 
-	this.object = object;
-	this.domElement = ( domElement !== undefined ) ? domElement : document;
+  // API
 
-	// API
+  this.enabled = true
 
-	this.enabled = true;
+  this.screen = { width: 0, height: 0, offsetLeft: 0, offsetTop: 0 }
+  this.radius = (this.screen.width + this.screen.height) / 4
 
-	this.screen = { width: 0, height: 0, offsetLeft: 0, offsetTop: 0 };
-	this.radius = ( this.screen.width + this.screen.height ) / 4;
+  this.rotateSpeed = 1.0
+  this.zoomSpeed = 1.2
+  this.panSpeed = 0.3
 
-	this.rotateSpeed = 1.0;
-	this.zoomSpeed = 1.2;
-	this.panSpeed = 0.3;
+  this.noRotate = false
+  this.noZoom = false
+  this.noPan = false
 
-	this.noRotate = false;
-	this.noZoom = false;
-	this.noPan = false;
+  this.staticMoving = false
+  this.dynamicDampingFactor = 0.2
 
-	this.staticMoving = false;
-	this.dynamicDampingFactor = 0.2;
+  this.minDistance = 0
+  this.maxDistance = Infinity
 
-	this.minDistance = 0;
-	this.maxDistance = Infinity;
+  this.keys = [
+    65, /* A */
+    83, /* S */
+    68 /* D */
+  ]
 
-	this.keys = [ 65 /*A*/, 83 /*S*/, 68 /*D*/ ];
+  // internals
 
-	// internals
+  this.target = new THREE.Vector3()
 
-	this.target = new THREE.Vector3();
+  var lastPosition = new THREE.Vector3()
 
-	var lastPosition = new THREE.Vector3();
+  var _state = STATE.NONE
+  var _prevState = STATE.NONE
 
-	var _state = STATE.NONE,
-	_prevState = STATE.NONE,
+  var _eye = new THREE.Vector3()
 
-	_eye = new THREE.Vector3(),
+  var _rotateStart = new THREE.Vector3()
+  var _rotateEnd = new THREE.Vector3()
 
-	_rotateStart = new THREE.Vector3(),
-	_rotateEnd = new THREE.Vector3(),
+  var _zoomStart = new THREE.Vector2()
+  var _zoomEnd = new THREE.Vector2()
 
-	_zoomStart = new THREE.Vector2(),
-	_zoomEnd = new THREE.Vector2(),
+  var _touchZoomDistanceStart = 0
+  var _touchZoomDistanceEnd = 0
 
-	_touchZoomDistanceStart = 0,
-	_touchZoomDistanceEnd = 0,
+  var _panStart = new THREE.Vector2()
+  var _panEnd = new THREE.Vector2()
 
-	_panStart = new THREE.Vector2(),
-	_panEnd = new THREE.Vector2();
+  // for reset
 
-	// for reset
+  this.target0 = this.target.clone()
+  this.position0 = this.object.position.clone()
+  this.up0 = this.object.up.clone()
 
-	this.target0 = this.target.clone();
-	this.position0 = this.object.position.clone();
-	this.up0 = this.object.up.clone();
+  // events
 
-	// events
+  var changeEvent = { type: 'change' }
 
-	var changeEvent = { type: 'change' };
+  // methods
 
+  this.handleResize = function () {
+    this.screen.width = window.innerWidth
+    this.screen.height = window.innerHeight
 
-	// methods
+    this.screen.offsetLeft = 0
+    this.screen.offsetTop = 0
 
-	this.handleResize = function () {
+    this.radius = (this.screen.width + this.screen.height) / 4
+  }
 
-		this.screen.width = window.innerWidth;
-		this.screen.height = window.innerHeight;
+  this.handleEvent = function (event) {
+    if (typeof this[ event.type ] === 'function') {
+      this[ event.type ](event)
+    }
+  }
 
-		this.screen.offsetLeft = 0;
-		this.screen.offsetTop = 0;
+  this.getMouseOnScreen = function (clientX, clientY) {
+    return new THREE.Vector2(
+      (clientX - _this.screen.offsetLeft) / _this.radius * 0.5,
+      (clientY - _this.screen.offsetTop) / _this.radius * 0.5
+    )
+  }
 
-		this.radius = ( this.screen.width + this.screen.height ) / 4;
+  this.getMouseProjectionOnBall = function (clientX, clientY) {
+    var mouseOnBall = new THREE.Vector3(
+      (clientX - _this.screen.width * 0.5 - _this.screen.offsetLeft) / _this.radius,
+      (_this.screen.height * 0.5 + _this.screen.offsetTop - clientY) / _this.radius,
+      0.0
+    )
 
-	};
+    var length = mouseOnBall.length()
+
+    if (length > 1.0) {
+      mouseOnBall.normalize()
+    } else {
+      mouseOnBall.z = Math.sqrt(1.0 - length * length)
+    }
+
+    _eye.copy(_this.object.position).sub(_this.target)
+
+    var projection = _this.object.up.clone().setLength(mouseOnBall.y)
+    projection.add(_this.object.up.clone().cross(_eye).setLength(mouseOnBall.x))
+    projection.add(_eye.setLength(mouseOnBall.z))
 
-	this.handleEvent = function ( event ) {
+    return projection
+  }
 
-		if ( typeof this[ event.type ] == 'function' ) {
+  this.rotateCamera = function () {
+    var angle = Math.acos(_rotateStart.dot(_rotateEnd) / _rotateStart.length() / _rotateEnd.length())
 
-			this[ event.type ]( event );
+    if (angle) {
+      var axis = (new THREE.Vector3()).crossVectors(_rotateStart, _rotateEnd).normalize()
+      var quaternion = new THREE.Quaternion()
 
-		}
+      angle *= _this.rotateSpeed
 
-	};
+      quaternion.setFromAxisAngle(axis, -angle)
 
-	this.getMouseOnScreen = function ( clientX, clientY ) {
+      _eye.applyQuaternion(quaternion)
+      _this.object.up.applyQuaternion(quaternion)
 
-		return new THREE.Vector2(
-			( clientX - _this.screen.offsetLeft ) / _this.radius * 0.5,
-			( clientY - _this.screen.offsetTop ) / _this.radius * 0.5
-		);
+      _rotateEnd.applyQuaternion(quaternion)
 
-	};
+      if (_this.staticMoving) {
+        _rotateStart.copy(_rotateEnd)
+      } else {
+        quaternion.setFromAxisAngle(axis, angle * (_this.dynamicDampingFactor - 1.0))
+        _rotateStart.applyQuaternion(quaternion)
+      }
+    }
+  }
 
-	this.getMouseProjectionOnBall = function ( clientX, clientY ) {
+  this.zoomCamera = function () {
+    if (_state === STATE.TOUCH_ZOOM) {
+      let factor = _touchZoomDistanceStart / _touchZoomDistanceEnd
+      _touchZoomDistanceStart = _touchZoomDistanceEnd
+      _eye.multiplyScalar(factor)
+    } else {
+      let factor = 1.0 + (_zoomEnd.y - _zoomStart.y) * _this.zoomSpeed
 
-		var mouseOnBall = new THREE.Vector3(
-			( clientX - _this.screen.width * 0.5 - _this.screen.offsetLeft ) / _this.radius,
-			( _this.screen.height * 0.5 + _this.screen.offsetTop - clientY ) / _this.radius,
-			0.0
-		);
+      if (factor !== 1.0 && factor > 0.0) {
+        _eye.multiplyScalar(factor)
 
-		var length = mouseOnBall.length();
+        if (_this.staticMoving) {
+          _zoomStart.copy(_zoomEnd)
+        } else {
+          _zoomStart.y += (_zoomEnd.y - _zoomStart.y) * this.dynamicDampingFactor
+        }
+      }
+    }
+  }
 
-		if ( length > 1.0 ) {
+  this.panCamera = function () {
+    var mouseChange = _panEnd.clone().sub(_panStart)
 
-			mouseOnBall.normalize();
+    if (mouseChange.lengthSq()) {
+      mouseChange.multiplyScalar(_eye.length() * _this.panSpeed)
 
-		} else {
+      var pan = _eye.clone().cross(_this.object.up).setLength(mouseChange.x)
+      pan.add(_this.object.up.clone().setLength(mouseChange.y))
 
-			mouseOnBall.z = Math.sqrt( 1.0 - length * length );
+      _this.object.position.add(pan)
+      _this.target.add(pan)
 
-		}
+      if (_this.staticMoving) {
+        _panStart = _panEnd
+      } else {
+        _panStart.add(mouseChange.subVectors(_panEnd, _panStart).multiplyScalar(_this.dynamicDampingFactor))
+      }
+    }
+  }
 
-		_eye.copy( _this.object.position ).sub( _this.target );
+  this.checkDistances = function () {
+    if (!_this.noZoom || !_this.noPan) {
+      if (_this.object.position.lengthSq() > _this.maxDistance * _this.maxDistance) {
+        _this.object.position.setLength(_this.maxDistance)
+      }
 
-		var projection = _this.object.up.clone().setLength( mouseOnBall.y );
-		projection.add( _this.object.up.clone().cross( _eye ).setLength( mouseOnBall.x ) );
-		projection.add( _eye.setLength( mouseOnBall.z ) );
+      if (_eye.lengthSq() < _this.minDistance * _this.minDistance) {
+        _this.object.position.addVectors(_this.target, _eye.setLength(_this.minDistance))
+      }
+    }
+  }
 
-		return projection;
+  this.update = function () {
+    _eye.subVectors(_this.object.position, _this.target)
 
-	};
+    if (!_this.noRotate) {
+      _this.rotateCamera()
+    }
 
-	this.rotateCamera = function () {
+    if (!_this.noZoom) {
+      _this.zoomCamera()
+    }
 
-		var angle = Math.acos( _rotateStart.dot( _rotateEnd ) / _rotateStart.length() / _rotateEnd.length() );
+    if (!_this.noPan) {
+      _this.panCamera()
+    }
 
-		if ( angle ) {
+    _this.object.position.addVectors(_this.target, _eye)
 
-			var axis = ( new THREE.Vector3() ).crossVectors( _rotateStart, _rotateEnd ).normalize(),
-				quaternion = new THREE.Quaternion();
+    _this.checkDistances()
 
-			angle *= _this.rotateSpeed;
+    _this.object.lookAt(_this.target)
 
-			quaternion.setFromAxisAngle( axis, -angle );
+    if (lastPosition.distanceToSquared(_this.object.position) > 0) {
+      // if(typeof (_this.dispatchEvent) === function) {
+      // _this.dispatchEvent( changeEvent );
+      // }
 
-			_eye.applyQuaternion( quaternion );
-			_this.object.up.applyQuaternion( quaternion );
+      lastPosition.copy(_this.object.position)
+    }
+  }
 
-			_rotateEnd.applyQuaternion( quaternion );
+  this.reset = function () {
+    _state = STATE.NONE
+    _prevState = STATE.NONE
 
-			if ( _this.staticMoving ) {
+    _this.target.copy(_this.target0)
+    _this.object.position.copy(_this.position0)
+    _this.object.up.copy(_this.up0)
 
-				_rotateStart.copy( _rotateEnd );
+    _eye.subVectors(_this.object.position, _this.target)
 
-			} else {
+    _this.object.lookAt(_this.target)
 
-				quaternion.setFromAxisAngle( axis, angle * ( _this.dynamicDampingFactor - 1.0 ) );
-				_rotateStart.applyQuaternion( quaternion );
+    _this.dispatchEvent(changeEvent)
 
-			}
+    lastPosition.copy(_this.object.position)
+  }
 
-		}
+  // listeners
 
-	};
+  function keydown (event) {
+    if (_this.enabled === false) return
 
-	this.zoomCamera = function () {
+    window.removeEventListener('keydown', keydown)
 
-		if ( _state === STATE.TOUCH_ZOOM ) {
+    _prevState = _state
 
-			var factor = _touchZoomDistanceStart / _touchZoomDistanceEnd;
-			_touchZoomDistanceStart = _touchZoomDistanceEnd;
-			_eye.multiplyScalar( factor );
+    if (_state !== STATE.NONE) {
 
-		} else {
+    } else if (event.keyCode === _this.keys[ STATE.ROTATE ] && !_this.noRotate) {
+      _state = STATE.ROTATE
+    } else if (event.keyCode === _this.keys[ STATE.ZOOM ] && !_this.noZoom) {
+      _state = STATE.ZOOM
+    } else if (event.keyCode === _this.keys[ STATE.PAN ] && !_this.noPan) {
+      _state = STATE.PAN
+    }
+  }
 
-			var factor = 1.0 + ( _zoomEnd.y - _zoomStart.y ) * _this.zoomSpeed;
+  function keyup (event) {
+    if (_this.enabled === false) return
 
-			if ( factor !== 1.0 && factor > 0.0 ) {
+    _state = _prevState
 
-				_eye.multiplyScalar( factor );
+    window.addEventListener('keydown', keydown, false)
+  }
 
-				if ( _this.staticMoving ) {
+  function mousedown (event) {
+    if (_this.enabled === false) return
 
-					_zoomStart.copy( _zoomEnd );
+    event.preventDefault()
+    event.stopPropagation()
 
-				} else {
+    if (_state === STATE.NONE) {
+      _state = event.button
+    }
 
-					_zoomStart.y += ( _zoomEnd.y - _zoomStart.y ) * this.dynamicDampingFactor;
+    if (_state === STATE.ROTATE && !_this.noRotate) {
+      _rotateStart = _rotateEnd = _this.getMouseProjectionOnBall(event.clientX, event.clientY)
+    } else if (_state === STATE.ZOOM && !_this.noZoom) {
+      _zoomStart = _zoomEnd = _this.getMouseOnScreen(event.clientX, event.clientY)
+    } else if (_state === STATE.PAN && !_this.noPan) {
+      _panStart = _panEnd = _this.getMouseOnScreen(event.clientX, event.clientY)
+    }
 
-				}
+    document.addEventListener('mousemove', mousemove, false)
+    document.addEventListener('mouseup', mouseup, false)
+  }
 
-			}
+  function mousemove (event) {
+    if (_this.enabled === false) return
 
-		}
+    event.preventDefault()
+    event.stopPropagation()
 
-	};
+    if (_state === STATE.ROTATE && !_this.noRotate) {
+      _rotateEnd = _this.getMouseProjectionOnBall(event.clientX, event.clientY)
+    } else if (_state === STATE.ZOOM && !_this.noZoom) {
+      _zoomEnd = _this.getMouseOnScreen(event.clientX, event.clientY)
+    } else if (_state === STATE.PAN && !_this.noPan) {
+      _panEnd = _this.getMouseOnScreen(event.clientX, event.clientY)
+    }
+  }
 
-	this.panCamera = function () {
+  function mouseup (event) {
+    if (_this.enabled === false) return
 
-		var mouseChange = _panEnd.clone().sub( _panStart );
+    event.preventDefault()
+    event.stopPropagation()
 
-		if ( mouseChange.lengthSq() ) {
+    _state = STATE.NONE
 
-			mouseChange.multiplyScalar( _eye.length() * _this.panSpeed );
+    document.removeEventListener('mousemove', mousemove)
+    document.removeEventListener('mouseup', mouseup)
+  }
 
-			var pan = _eye.clone().cross( _this.object.up ).setLength( mouseChange.x );
-			pan.add( _this.object.up.clone().setLength( mouseChange.y ) );
+  function mousewheel (event) {
+    if (_this.enabled === false) return
 
-			_this.object.position.add( pan );
-			_this.target.add( pan );
+    event.preventDefault()
+    event.stopPropagation()
 
-			if ( _this.staticMoving ) {
+    var delta = 0
 
-				_panStart = _panEnd;
+    if (event.wheelDelta) { // WebKit / Opera / Explorer 9
+      delta = event.wheelDelta / 40
+    } else if (event.detail) { // Firefox
+      delta = -event.detail / 3
+    }
 
-			} else {
+    _zoomStart.y += (1 / delta) * 0.05
+  }
 
-				_panStart.add( mouseChange.subVectors( _panEnd, _panStart ).multiplyScalar( _this.dynamicDampingFactor ) );
+  function touchstart (event) {
+    if (_this.enabled === false) return
 
-			}
+    switch (event.touches.length) {
+      case 1:
+        _state = STATE.TOUCH_ROTATE
+        _rotateStart = _rotateEnd = _this.getMouseProjectionOnBall(event.touches[ 0 ].pageX, event.touches[ 0 ].pageY)
+        break
 
-		}
+      case 2:
+        _state = STATE.TOUCH_ZOOM
+        var dx = event.touches[ 0 ].pageX - event.touches[ 1 ].pageX
+        var dy = event.touches[ 0 ].pageY - event.touches[ 1 ].pageY
+        _touchZoomDistanceEnd = _touchZoomDistanceStart = Math.sqrt(dx * dx + dy * dy)
+        break
 
-	};
+      case 3:
+        _state = STATE.TOUCH_PAN
+        _panStart = _panEnd = _this.getMouseOnScreen(event.touches[ 0 ].pageX, event.touches[ 0 ].pageY)
+        break
 
-	this.checkDistances = function () {
+      default:
+        _state = STATE.NONE
+    }
+  }
 
-		if ( !_this.noZoom || !_this.noPan ) {
+  function touchmove (event) {
+    if (_this.enabled === false) return
 
-			if ( _this.object.position.lengthSq() > _this.maxDistance * _this.maxDistance ) {
+    event.preventDefault()
+    event.stopPropagation()
 
-				_this.object.position.setLength( _this.maxDistance );
+    switch (event.touches.length) {
+      case 1:
+        _rotateEnd = _this.getMouseProjectionOnBall(event.touches[ 0 ].pageX, event.touches[ 0 ].pageY)
+        break
 
-			}
+      case 2:
+        var dx = event.touches[ 0 ].pageX - event.touches[ 1 ].pageX
+        var dy = event.touches[ 0 ].pageY - event.touches[ 1 ].pageY
+        _touchZoomDistanceEnd = Math.sqrt(dx * dx + dy * dy)
+        break
 
-			if ( _eye.lengthSq() < _this.minDistance * _this.minDistance ) {
+      case 3:
+        _panEnd = _this.getMouseOnScreen(event.touches[ 0 ].pageX, event.touches[ 0 ].pageY)
+        break
 
-				_this.object.position.addVectors( _this.target, _eye.setLength( _this.minDistance ) );
+      default:
+        _state = STATE.NONE
+    }
+  }
 
-			}
+  function touchend (event) {
+    if (_this.enabled === false) return
 
-		}
+    switch (event.touches.length) {
+      case 1:
+        _rotateStart = _rotateEnd = _this.getMouseProjectionOnBall(event.touches[ 0 ].pageX, event.touches[ 0 ].pageY)
+        break
 
-	};
+      case 2:
+        _touchZoomDistanceStart = _touchZoomDistanceEnd = 0
+        break
 
-	this.update = function () {
+      case 3:
+        _panStart = _panEnd = _this.getMouseOnScreen(event.touches[ 0 ].pageX, event.touches[ 0 ].pageY)
+        break
+    }
 
-		_eye.subVectors( _this.object.position, _this.target );
+    _state = STATE.NONE
+  }
 
-		if ( !_this.noRotate ) {
+  this.domElement.addEventListener('contextmenu', function (event) { event.preventDefault() }, false)
 
-			_this.rotateCamera();
+  this.domElement.addEventListener('mousedown', mousedown, false)
 
-		}
+  this.domElement.addEventListener('mousewheel', mousewheel, false)
+  this.domElement.addEventListener('DOMMouseScroll', mousewheel, false) // firefox
 
-		if ( !_this.noZoom ) {
+  this.domElement.addEventListener('touchstart', touchstart, false)
+  this.domElement.addEventListener('touchend', touchend, false)
+  this.domElement.addEventListener('touchmove', touchmove, false)
 
-			_this.zoomCamera();
+  window.addEventListener('keydown', keydown, false)
+  window.addEventListener('keyup', keyup, false)
 
-		}
-
-		if ( !_this.noPan ) {
-
-			_this.panCamera();
-
-		}
-
-		_this.object.position.addVectors( _this.target, _eye );
-
-		_this.checkDistances();
-
-		_this.object.lookAt( _this.target );
-
-		if ( lastPosition.distanceToSquared( _this.object.position ) > 0 ) {
-
-			// if(typeof (_this.dispatchEvent) === function) {
-				// _this.dispatchEvent( changeEvent );
-			// }
-
-			lastPosition.copy( _this.object.position );
-
-		}
-
-	};
-
-	this.reset = function () {
-
-		_state = STATE.NONE;
-		_prevState = STATE.NONE;
-
-		_this.target.copy( _this.target0 );
-		_this.object.position.copy( _this.position0 );
-		_this.object.up.copy( _this.up0 );
-
-		_eye.subVectors( _this.object.position, _this.target );
-
-		_this.object.lookAt( _this.target );
-
-		_this.dispatchEvent( changeEvent );
-
-		lastPosition.copy( _this.object.position );
-
-	};
-
-	// listeners
-
-	function keydown( event ) {
-
-		if ( _this.enabled === false ) return;
-
-		window.removeEventListener( 'keydown', keydown );
-
-		_prevState = _state;
-
-		if ( _state !== STATE.NONE ) {
-
-			return;
-
-		} else if ( event.keyCode === _this.keys[ STATE.ROTATE ] && !_this.noRotate ) {
-
-			_state = STATE.ROTATE;
-
-		} else if ( event.keyCode === _this.keys[ STATE.ZOOM ] && !_this.noZoom ) {
-
-			_state = STATE.ZOOM;
-
-		} else if ( event.keyCode === _this.keys[ STATE.PAN ] && !_this.noPan ) {
-
-			_state = STATE.PAN;
-
-		}
-
-	}
-
-	function keyup( event ) {
-
-		if ( _this.enabled === false ) return;
-
-		_state = _prevState;
-
-		window.addEventListener( 'keydown', keydown, false );
-
-	}
-
-	function mousedown( event ) {
-
-		if ( _this.enabled === false ) return;
-
-		event.preventDefault();
-		event.stopPropagation();
-
-		if ( _state === STATE.NONE ) {
-
-			_state = event.button;
-
-		}
-
-		if ( _state === STATE.ROTATE && !_this.noRotate ) {
-
-			_rotateStart = _rotateEnd = _this.getMouseProjectionOnBall( event.clientX, event.clientY );
-
-		} else if ( _state === STATE.ZOOM && !_this.noZoom ) {
-
-			_zoomStart = _zoomEnd = _this.getMouseOnScreen( event.clientX, event.clientY );
-
-		} else if ( _state === STATE.PAN && !_this.noPan ) {
-
-			_panStart = _panEnd = _this.getMouseOnScreen( event.clientX, event.clientY );
-
-		}
-
-		document.addEventListener( 'mousemove', mousemove, false );
-		document.addEventListener( 'mouseup', mouseup, false );
-
-	}
-
-	function mousemove( event ) {
-
-		if ( _this.enabled === false ) return;
-
-		event.preventDefault();
-		event.stopPropagation();
-
-		if ( _state === STATE.ROTATE && !_this.noRotate ) {
-
-			_rotateEnd = _this.getMouseProjectionOnBall( event.clientX, event.clientY );
-
-		} else if ( _state === STATE.ZOOM && !_this.noZoom ) {
-
-			_zoomEnd = _this.getMouseOnScreen( event.clientX, event.clientY );
-
-		} else if ( _state === STATE.PAN && !_this.noPan ) {
-
-			_panEnd = _this.getMouseOnScreen( event.clientX, event.clientY );
-
-		}
-
-	}
-
-	function mouseup( event ) {
-
-		if ( _this.enabled === false ) return;
-
-		event.preventDefault();
-		event.stopPropagation();
-
-		_state = STATE.NONE;
-
-		document.removeEventListener( 'mousemove', mousemove );
-		document.removeEventListener( 'mouseup', mouseup );
-
-	}
-
-	function mousewheel( event ) {
-
-		if ( _this.enabled === false ) return;
-
-		event.preventDefault();
-		event.stopPropagation();
-
-		var delta = 0;
-
-		if ( event.wheelDelta ) { // WebKit / Opera / Explorer 9
-
-			delta = event.wheelDelta / 40;
-
-		} else if ( event.detail ) { // Firefox
-
-			delta = - event.detail / 3;
-
-		}
-
-		_zoomStart.y += ( 1 / delta ) * 0.05;
-
-	}
-
-	function touchstart( event ) {
-
-		if ( _this.enabled === false ) return;
-
-		switch ( event.touches.length ) {
-
-			case 1:
-				_state = STATE.TOUCH_ROTATE;
-				_rotateStart = _rotateEnd = _this.getMouseProjectionOnBall( event.touches[ 0 ].pageX, event.touches[ 0 ].pageY );
-				break;
-
-			case 2:
-				_state = STATE.TOUCH_ZOOM;
-				var dx = event.touches[ 0 ].pageX - event.touches[ 1 ].pageX;
-				var dy = event.touches[ 0 ].pageY - event.touches[ 1 ].pageY;
-				_touchZoomDistanceEnd = _touchZoomDistanceStart = Math.sqrt( dx * dx + dy * dy );
-				break;
-
-			case 3:
-				_state = STATE.TOUCH_PAN;
-				_panStart = _panEnd = _this.getMouseOnScreen( event.touches[ 0 ].pageX, event.touches[ 0 ].pageY );
-				break;
-
-			default:
-				_state = STATE.NONE;
-
-		}
-
-	}
-
-	function touchmove( event ) {
-
-		if ( _this.enabled === false ) return;
-
-		event.preventDefault();
-		event.stopPropagation();
-
-		switch ( event.touches.length ) {
-
-			case 1:
-				_rotateEnd = _this.getMouseProjectionOnBall( event.touches[ 0 ].pageX, event.touches[ 0 ].pageY );
-				break;
-
-			case 2:
-				var dx = event.touches[ 0 ].pageX - event.touches[ 1 ].pageX;
-				var dy = event.touches[ 0 ].pageY - event.touches[ 1 ].pageY;
-				_touchZoomDistanceEnd = Math.sqrt( dx * dx + dy * dy )
-				break;
-
-			case 3:
-				_panEnd = _this.getMouseOnScreen( event.touches[ 0 ].pageX, event.touches[ 0 ].pageY );
-				break;
-
-			default:
-				_state = STATE.NONE;
-
-		}
-
-	}
-
-	function touchend( event ) {
-
-		if ( _this.enabled === false ) return;
-
-		switch ( event.touches.length ) {
-
-			case 1:
-				_rotateStart = _rotateEnd = _this.getMouseProjectionOnBall( event.touches[ 0 ].pageX, event.touches[ 0 ].pageY );
-				break;
-
-			case 2:
-				_touchZoomDistanceStart = _touchZoomDistanceEnd = 0;
-				break;
-
-			case 3:
-				_panStart = _panEnd = _this.getMouseOnScreen( event.touches[ 0 ].pageX, event.touches[ 0 ].pageY );
-				break;
-
-		}
-
-		_state = STATE.NONE;
-
-	}
-
-	this.domElement.addEventListener( 'contextmenu', function ( event ) { event.preventDefault(); }, false );
-
-	this.domElement.addEventListener( 'mousedown', mousedown, false );
-
-	this.domElement.addEventListener( 'mousewheel', mousewheel, false );
-	this.domElement.addEventListener( 'DOMMouseScroll', mousewheel, false ); // firefox
-
-	this.domElement.addEventListener( 'touchstart', touchstart, false );
-	this.domElement.addEventListener( 'touchend', touchend, false );
-	this.domElement.addEventListener( 'touchmove', touchmove, false );
-
-	window.addEventListener( 'keydown', keydown, false );
-	window.addEventListener( 'keyup', keyup, false );
-
-	this.handleResize();
+  this.handleResize()
 };
