@@ -1,11 +1,9 @@
-import api from '@/utils/api'
+import { service as api } from '@/utils/api'
 
-export function getInfo(token) {
+export function getInfo (token) {
   return api({
     url: '/login/testGetInfo',
     method: 'get',
-    params: {token}
+    params: { token }
   })
 }
-
-
