@@ -32,14 +32,14 @@ const app = {
     }
   },
   actions: {
-    ToggleSideBar({ commit , rootState}) {
+    ToggleSideBar ({ commit, rootState }) {
       console.log(rootState.count)
       commit('TOGGLE_SIDEBAR')
     },
-    addVisitedViews({ commit }, view) {
+    addVisitedViews ({ commit }, view) {
       commit('ADD_VISITED_VIEWS', view)
     },
-    delVisitedViews({ commit, state }, view) {
+    delVisitedViews ({ commit, state }, view) {
       return new Promise((resolve) => {
         commit('DEL_VISITED_VIEWS', view)
         resolve([...state.visitedViews])
