@@ -1,6 +1,6 @@
 const counter = {
   state: {
-    count: 0,
+    count: 0
   },
   getters: {
     COUNT: state => {
@@ -22,20 +22,38 @@ const counter = {
     }
   },
   actions: {
-    incrementNum: ({commit, state}, num) => {
+    incrementNum: ({ commit, state }, num) => {
       commit('NUM', num)
     },
-    payLoad: ({commit, state}, pay) => {
+    payLoad: ({ commit, state }, pay) => {
       commit('PAYLOAD', pay)
     },
     async ad (context) {
-      context.commit('INCREMENT')
+      return new Promise((resolve, reject) => {
+        setTimeout(() => {
+          context.commit('INCREMENT')
+          resolve(context.getters.COUNT)
+        }, 2000)
+      })
     },
     async dd (context) {
-      context.commit('DECREMENT')
+      return new Promise((resolve, reject) => {
+        setTimeout(() => {
+          context.commit('DECREMENT')
+          resolve(context.getters.COUNT)
+        }, 3000)
+      })
     },
     async bw (context) {
       await context.dispatch('ad')
+    },
+    async y1 (context) {
+      let a = await context.dispatch('ad')
+      if (a > 10) {
+        await context.dispatch('dd')
+      } else {
+        await context.dispatch('ad')
+      }
     }
   }
 }
