@@ -2,10 +2,10 @@
   <div style="background:#f0f2f5;margin-top: -20px;">
     <div class="wscn-http404">
       <div class="pic-404">
-        <img class="pic-404__parent" :src="img_404" alt="404">
-        <img class="pic-404__child left" :src="img_404_cloud" alt="404">
-        <img class="pic-404__child mid" :src="img_404_cloud" alt="404">
-        <img class="pic-404__child right" :src="img_404_cloud" alt="404">
+        <img class="pic-404__parent" :src="img404" alt="404">
+        <img class="pic-404__child left" :src="img404cloud" alt="404">
+        <img class="pic-404__child mid" :src="img404cloud" alt="404">
+        <img class="pic-404__child right" :src="img404cloud" alt="404">
       </div>
       <div class="bullshit">
         <div class="bullshit__oops">OOPS!</div>
@@ -18,22 +18,22 @@
 </template>
 
 <script>
- import img_404 from '@/assets/404_images/404.png'
- import img_404_cloud from '@/assets/404_images/404_cloud.png'
+import img404 from '@/assets/404_images/404.png'
+import img404cloud from '@/assets/404_images/404_cloud.png'
 
- export default {
-   data() {
-     return {
-       img_404,
-       img_404_cloud
-     }
-   },
-   computed: {
-     message() {
-       return '页面找不到了......'
-     }
-   }
- }
+export default {
+  data () {
+    return {
+      img404,
+      img404cloud
+    }
+  },
+  computed: {
+    message () {
+      return '页面找不到了......'
+    }
+  }
+}
 </script>
 
 <style rel="stylesheet/scss" lang="scss" scoped>
