@@ -1,6 +1,6 @@
 <template>
   <section class="app-main">
-    <transition name="fade" mode="out-in">
+    <transition name="fade" mode="out-in" enter-active-class="animated slideInLeft">
       <!-- <router-view :key="key"></router-view> -->
       <router-view></router-view>
     </transition>
@@ -10,16 +10,22 @@
 <script>
 export default {
   name: 'AppMain',
+  data () {
+    return {
+    }
+  },
   computed: {
     // key() {
     //   return this.$route.name !== undefined ? this.$route.name + +new Date() : this.$route + +new Date()
     // }
+  },
+  methods: {
   }
 }
 </script>
 
 <style>
-.app-main {
-  padding: 20px;
-}
+  .app-main {
+    padding: 20px;
+  }
 </style>
