@@ -1,32 +1,34 @@
 <template>
   <div class="menu-wrapper">
-    <template v-for="item in routes" v-if="!item.hidden&&item.children">
+    <template v-for="item in routes">
+      <template v-if="!item.hidden&&item.children">
+        <router-link v-if="item.children.length===1 && !item.children[0].children" :to="item.path+'/'+item.children[0].path" :key="item.children[0].name">
+          <el-menu-item :index="item.path+'/'+item.children[0].path" class='submenu-title-noDropdown'>
+            <svg-icon v-if="item.children[0].meta&&item.children[0].meta.icon" :icon-class="item.children[0].meta.icon"></svg-icon>
+            <span v-if="item.children[0].meta&&item.children[0].meta.title">{{item.children[0].meta.title}}</span>
+          </el-menu-item>
+        </router-link>
 
-      <router-link v-if="item.children.length===1 && !item.children[0].children" :to="item.path+'/'+item.children[0].path" :key="item.children[0].name">
-        <el-menu-item :index="item.path+'/'+item.children[0].path" class='submenu-title-noDropdown'>
-          <svg-icon v-if="item.children[0].meta&&item.children[0].meta.icon" :icon-class="item.children[0].meta.icon"></svg-icon>
-          <span v-if="item.children[0].meta&&item.children[0].meta.title">{{item.children[0].meta.title}}</span>
-        </el-menu-item>
-      </router-link>
+        <el-submenu v-else :index="item.name||item.path" :key="item.name">
+          <template slot="title">
+            <svg-icon v-if="item.meta&&item.meta.icon" :icon-class="item.meta.icon"></svg-icon>
+            <span v-if="item.meta&&item.meta.title">{{item.meta.title}}</span>
+          </template>
 
-      <el-submenu v-else :index="item.name||item.path" :key="item.name">
-        <template slot="title">
-          <svg-icon v-if="item.meta&&item.meta.icon" :icon-class="item.meta.icon"></svg-icon>
-          <span v-if="item.meta&&item.meta.title">{{item.meta.title}}</span>
-        </template>
+          <template v-for="child in item.children">
+            <template v-if="!child.hidden">
+              <sidebar-item class="nest-menu" v-if="child.children&&child.children.length>0" :selfroute="item.path" :routes="[child]" :key="child.path"></sidebar-item>
 
-        <template v-for="child in item.children" v-if="!child.hidden">
-          <sidebar-item class="nest-menu" v-if="child.children&&child.children.length>0" :routes="[child]" :key="child.path"></sidebar-item>
-
-          <router-link v-else :to="item.path+'/'+child.path" :key="child.name">
-            <el-menu-item :index="item.path+'/'+child.path">
-              <svg-icon v-if="child.meta&&child.meta.icon" :icon-class="child.meta.icon"></svg-icon>
-              <span v-if="child.meta&&child.meta.title">{{child.meta.title}}</span>
-            </el-menu-item>
-          </router-link>
-        </template>
-      </el-submenu>
-
+              <router-link v-else :to="selfroute+(selfroute == '' ? '' : '/')+item.path+'/'+child.path" :key="child.name">
+                <el-menu-item :index="selfroute+(selfroute == '' ? '' : '/')+item.path+'/'+child.path">
+                  <svg-icon v-if="child.meta&&child.meta.icon" :icon-class="child.meta.icon"></svg-icon>
+                  <span v-if="child.meta&&child.meta.title">{{child.meta.title}}</span>
+                </el-menu-item>
+              </router-link>
+            </template>
+          </template>
+        </el-submenu>
+      </template>
     </template>
   </div>
 </template>
@@ -37,6 +39,9 @@ export default {
   props: {
     routes: {
       type: Array
+    },
+    selfroute: {
+      type: String
     }
   }
 }
