@@ -1,26 +1,26 @@
 <template>
   <scroll-bar>
     <el-menu mode="vertical" unique-opened :default-active="$route.path" :collapse="isCollapse"
-             background-color="#8DA79C" text-color="black" active-text-color="#eee">
-      <sidebar-item :routes="permission_routers"></sidebar-item>
+             background-color="white" text-color="black" active-text-color="#66B3FF">
+      <sidebar-item :routes="permission_routers" selfroute=""></sidebar-item>
     </el-menu>
   </scroll-bar>
 </template>
 <script>
-  import {mapGetters} from 'vuex'
-  import SidebarItem from './SidebarItem'
-  import ScrollBar from '@/components/ScrollBar'
+import { mapGetters } from 'vuex'
+import SidebarItem from './SidebarItem'
+import ScrollBar from '@/components/ScrollBar'
 
-  export default {
-    components: {SidebarItem, ScrollBar},
-    computed: {
-      ...mapGetters([
-        'permission_routers',
-        'sidebar'
-      ]),
-      isCollapse() {
-        return !this.sidebar.opened
-      }
+export default {
+  components: { SidebarItem, ScrollBar },
+  computed: {
+    ...mapGetters([
+      'permission_routers',
+      'sidebar'
+    ]),
+    isCollapse () {
+      return !this.sidebar.opened
     }
   }
+}
 </script>
