@@ -4,7 +4,7 @@
     <el-form autoComplete="on" :model="loginForm" :rules="loginRules" ref="loginForm" label-position="left"
              label-width="0px"
              class="card-box login-form">
-      <h3 class="title">vue学习管理系统</h3>
+      <h3 class="title">京东方(重庆)智造三车间</h3>
       <h5 class="title-side">@苏州鸿然</h5>
       <el-form-item prop="username">
         <span class="svg-container svg-container_login">
@@ -29,236 +29,235 @@
 </template>
 
 <script>
-  import * as THREE from 'three'
-  // import calc from '@/api/haha.js'
-  import TrackballControls from '@/api/TrackballControls.js'
-  import mountain from '@/images/mountain.jpg'
-  import ta from '@/images/ta.png'
-  export default {
-    name: 'login',
-    data() {
-      return {
-        loginForm: {
-          username: 'admin',
-          password: '123456'
-        },
-        loginRules: {
-          username: [{required: true, trigger: 'blur', message: "请输入用户名"}],
-          password: [{required: true, trigger: 'blur', message: "请输入密码"}]
-        },
-        loading: false,
-
-        camera: null,
-        scene: null,
-        renderer: null,
-        helper: null,
-        mesh: [],
-        width: window.innerWidth,
-        height: window.innerHeight
-
-      }
-    },
-    methods: {
-      onFocus(a) {
-        // var pFocus;
-        // if(a === 1){
-        //   pFocus = document.getElementById("user-name");
-        // } else if (a === 2){
-        //   pFocus = document.getElementById("user-password");
-        // }
-        //
-        // pFocus.focus();
-        // pFocus.select();
+import * as THREE from 'three'
+// import calc from '@/api/haha.js'
+import TrackballControls from '@/api/TrackballControls.js'
+import mountain from '@/images/mountain.jpg'
+import ta from '@/images/ta.png'
+export default {
+  name: 'login',
+  data () {
+    return {
+      loginForm: {
+        username: 'admin',
+        password: '123456'
       },
-      handleLogin() {
-        this.$refs.loginForm.validate(valid => {
-          if (valid) {
-            this.loading = true
-            this.$store.dispatch('Login', this.loginForm).then(data => {
-              this.loading = false
-              if ("success" === data.result) {
-                this.$router.push({path: '/'})
-              } else {
-                this.$message.error("账号/密码错误");
-              }
-            }).catch(() => {
-              this.loading = false
-            })
-          } else {
-            return false
-          }
-        })
+      loginRules: {
+        username: [{ required: true, trigger: 'blur', message: '请输入用户名' }],
+        password: [{ required: true, trigger: 'blur', message: '请输入密码' }]
       },
-      onWindowResize() {
-        this.width = window.innerWidth;
-        this.height = window.innerHeight;
-        this.camera.aspect = (this.width) / (this.height);
-        this.camera.updateProjectionMatrix();
-        this.renderer.setSize( this.width, this.height );
-        this.controls.handleResize();
-      },
-      init: function() {
-        let container = document.getElementById('container');
-        this.renderer = new THREE.WebGLRenderer({antialias: true});
-        this.renderer.setSize(this.width, this.height);
-        container.appendChild(this.renderer.domElement);
+      loading: false,
 
-        this.camera = new THREE.PerspectiveCamera(20, container.clientWidth/(container.clientHeight), 0.01, 10);
-        // this.camera.position.set(-2,-4,4);
-        this.camera.position.set(-4,3,2);
-        this.camera.lookAt(new THREE.Vector3(0, 0, 0));
+      camera: null,
+      scene: null,
+      renderer: null,
+      helper: null,
+      mesh: [],
+      width: window.innerWidth,
+      height: window.innerHeight
 
-        this.controls = new TrackballControls( this.camera,  this.renderer.domElement);
-        this.controls.rotateSpeed = 5.0;
-        this.controls.zoomSpeed = 5;
-        this.controls.panSpeed = 2;
-        this.controls.noZoom = false;
-        this.controls.noPan = false;
-        this.controls.staticMoving = false;
-        this.controls.dynamicDampingFactor = 0.3;
-
-        this.scene = new THREE.Scene();
-        this.scene.add(this.camera);
-
-        // var helper = new THREE.GridHelper( 50, 50 , 0x0000ff, 0x808080);
-        // this.scene.add( helper );
-
-        // this.helper = new THREE.GridHelper( 50, 50 , 0x0000ff, 0x808080);
-        // this.helper.rotation.x = Math.PI/2;
-        // this.scene.add( this.helper );
-
-        //let geometry = new THREE.BoxGeometry(0.2, 0.2, 0.2);
-        let geometry =  new THREE.SphereBufferGeometry( 0.13, 132, 116 );
-        //let texture = THREE.ImageUtils.loadTexture("src/images/mountain.jpg",null,function(t) { });
-        let texture = new THREE.TextureLoader().load(mountain);
-        //let material = new THREE.MeshNormalMaterial();
-        //let material = new THREE.MeshBasicMaterial({map:texture});
-        //let material = new THREE.MeshLambertMaterial( { color: 0xdddddd } );
-        //let material = new THREE.MeshBasicMaterial( { color: 0xffaacc, transparent: true, blending: THREE.AdditiveBlending } );
-        let material = new THREE.MeshPhongMaterial( { color: 0xdddddd, specular: 0xcccccc, shininess: 3, flatShading: true } )
-
-        // this.mesh = new THREE.Mesh(geometry, material);
-        // //this.mesh.position.y = 0.2;
-        // this.scene.add(this.mesh);
-
-        //let texture = new THREE.Texture( this.generateTexture() );
-
-        for(var i=-8; i<=8; i++) {
-          for(var j=-8; j<=8; j++) {
-            //for(var k=-2; k<=2; k++) {
-              if( i===1 && j===1) {
-                let geometry =  new THREE.SphereBufferGeometry( 0.23, 32, 16 );
-                let material = new THREE.MeshLambertMaterial( { map: texture, transparent: true } );
-                let mesh = new THREE.Mesh(geometry, material);
-                mesh.position.x = i*0.8;
-                mesh.position.y = j*0.8;
-                this.scene.add(mesh);
-                this.mesh.push(mesh);
-              } else if(i===0 && j===1) {
-                //let texture = THREE.ImageUtils.loadTexture("src/images/ta.png",null,function(t) { });
-                let texture = new THREE.TextureLoader().load(ta);
-                let geometry =  new THREE.SphereBufferGeometry( 0.23, 32, 16 );
-                let material = new THREE.MeshLambertMaterial( { map: texture, transparent: true } );
-                let mesh = new THREE.Mesh(geometry, material);
-                mesh.position.x = i*0.8;
-                mesh.position.y = j*0.8;
-                this.scene.add(mesh);
-                this.mesh.push(mesh);
-              } else if(i===1 && j===0) {
-                let geometry =  new THREE.SphereBufferGeometry( 0.23, 15, 15 );
-                let material = new THREE.MeshBasicMaterial( { color: 0xffaa00, wireframe: true } );
-                let mesh = new THREE.Mesh(geometry, material);
-                mesh.position.x = i*0.8;
-                mesh.position.y = j*0.8;
-                this.scene.add(mesh);
-                this.mesh.push(mesh);
-              } else if(i===-1 && j===1) {
-                let geometry =  new THREE.SphereBufferGeometry( 0.23, 35, 35 );
-                let material = new THREE.MeshPhongMaterial( { color: 0x000000, specular: 0x666666, emissive: 0x333333, shininess: 10, opacity: 0.9, transparent: true } );
-                let mesh = new THREE.Mesh(geometry, material);
-                mesh.position.x = i*0.8;
-                mesh.position.y = j*0.8;
-                this.scene.add(mesh);
-                this.mesh.push(mesh);
-              } else if(i===-1 && j===-1){
-                // shadow
-                var canvas = document.createElement( 'canvas' );
-                canvas.width = 2;
-                canvas.height = 2;
-                var context = canvas.getContext( '2d' );
-                var gradient = context.createRadialGradient( canvas.width / 2, canvas.height / 2, 0, canvas.width / 2, canvas.height / 2, canvas.width / 2 );
-                gradient.addColorStop( 0.1, 'rgba(210,210,210,1)' );
-                gradient.addColorStop( 1, 'rgba(255,255,255,1)' );
-                context.fillStyle = gradient;
-                context.fillRect( 0, 0, canvas.width, canvas.height );
-                var shadowTexture = new THREE.CanvasTexture( canvas );
-
-                let geometry =  new THREE.SphereBufferGeometry( 0.23, 32, 16 );
-                let material = new THREE.MeshLambertMaterial( { map: shadowTexture, transparent: true } );
-                let mesh = new THREE.Mesh(geometry, material);
-                mesh.position.x = i*0.8;
-                mesh.position.y = j*0.8;
-                this.scene.add(mesh);
-                this.mesh.push(mesh);
-              }
-              else {
-                let mesh = new THREE.Mesh(geometry, material);
-                mesh.position.x = i*0.8;
-                mesh.position.y = j*0.8;
-                mesh.position.z = Math.random();
-                //mesh.position.z = k*1;
-                this.scene.add(mesh);
-                if(i===0 && j===0) {
-
-                  this.mesh.push(mesh);
-                }
-                this.mesh.push(mesh);
-              }
-           //}
-          }
-        }
-
-        // Lights
-        this.scene.add( new THREE.AmbientLight( 0x222222 ) );
-        var directionalLight = new THREE.DirectionalLight( 0xaaaaaa, 0.125 );
-        directionalLight.position.x = Math.random() - 0.5;
-        directionalLight.position.y = Math.random() - 0.5;
-        directionalLight.position.z = Math.random() - 0.5;
-        directionalLight.position.normalize();
-        this.scene.add( directionalLight );
-        this.pointLight = new THREE.PointLight( 0xaaaaaa, 1 );
-        this.pointLight.position.z = 5;
-        this.scene.add( this.pointLight );
-
-        window.addEventListener( 'resize', this.onWindowResize, false );
-      },
-      animate: function() {
-
-        var timer = 0.0001 * Date.now();
-        this.camera.position.x = Math.cos( timer ) * 2;
-        //this.camera.position.y = Math.sin( timer ) * 2;
-        this.camera.position.z = Math.sin( timer ) * 2;
-        this.camera.lookAt( this.scene.position );
-        for ( var i = 0, l = this.mesh.length; i < l; i ++ ) {
-          var mesh = this.mesh[ i ];
-          mesh.rotation.x += 0.01;
-          mesh.rotation.y += 0.005;
-        }
-
-        this.pointLight.position.x = Math.sin( timer * 7 ) * 300;
-        this.pointLight.position.y = Math.cos( timer * 5 ) * 400;
-        this.pointLight.position.z = Math.cos( timer * 3 ) * 300;
-
-        requestAnimationFrame(this.animate);
-        this.controls.update();
-        this.renderer.render(this.scene, this.camera);
-      }
-    },
-    mounted() {
-      this.init()
-      this.animate()
     }
+  },
+  methods: {
+    onFocus (a) {
+      // var pFocus;
+      // if(a === 1){
+      //   pFocus = document.getElementById("user-name");
+      // } else if (a === 2){
+      //   pFocus = document.getElementById("user-password");
+      // }
+      //
+      // pFocus.focus();
+      // pFocus.select();
+    },
+    handleLogin () {
+      this.$refs.loginForm.validate(valid => {
+        if (valid) {
+          this.loading = true
+          this.$store.dispatch('Login', this.loginForm).then(data => {
+            this.loading = false
+            if (data.result === 'success') {
+              this.$router.push({ path: '/' })
+            } else if (data.result === 'fail') {
+              this.$message.error('账号/密码错误')
+            } else {
+              this.$message.error(data.result)
+            }
+          }).catch(() => {
+            this.loading = false
+          })
+        } else {
+          return false
+        }
+      })
+    },
+    onWindowResize () {
+      this.width = window.innerWidth
+      this.height = window.innerHeight
+      this.camera.aspect = (this.width) / (this.height)
+      this.camera.updateProjectionMatrix()
+      this.renderer.setSize(this.width, this.height)
+      this.controls.handleResize()
+    },
+    init: function () {
+      let container = document.getElementById('container')
+      this.renderer = new THREE.WebGLRenderer({ antialias: true })
+      this.renderer.setSize(this.width, this.height)
+      container.appendChild(this.renderer.domElement)
+
+      this.camera = new THREE.PerspectiveCamera(20, container.clientWidth / (container.clientHeight), 0.01, 10)
+      // this.camera.position.set(-2,-4,4);
+      this.camera.position.set(-4, 3, 2)
+      this.camera.lookAt(new THREE.Vector3(0, 0, 0))
+
+      this.controls = new TrackballControls(this.camera, this.renderer.domElement)
+      this.controls.rotateSpeed = 5.0
+      this.controls.zoomSpeed = 5
+      this.controls.panSpeed = 2
+      this.controls.noZoom = false
+      this.controls.noPan = false
+      this.controls.staticMoving = false
+      this.controls.dynamicDampingFactor = 0.3
+
+      this.scene = new THREE.Scene()
+      this.scene.add(this.camera)
+
+      // var helper = new THREE.GridHelper( 50, 50 , 0x0000ff, 0x808080);
+      // this.scene.add( helper );
+
+      // this.helper = new THREE.GridHelper( 50, 50 , 0x0000ff, 0x808080);
+      // this.helper.rotation.x = Math.PI/2;
+      // this.scene.add( this.helper );
+
+      // let geometry = new THREE.BoxGeometry(0.2, 0.2, 0.2);
+      let geometry = new THREE.SphereBufferGeometry(0.13, 132, 116)
+      // let texture = THREE.ImageUtils.loadTexture("src/images/mountain.jpg",null,function(t) { });
+      let texture = new THREE.TextureLoader().load(mountain)
+      // let material = new THREE.MeshNormalMaterial();
+      // let material = new THREE.MeshBasicMaterial({map:texture});
+      // let material = new THREE.MeshLambertMaterial( { color: 0xdddddd } );
+      // let material = new THREE.MeshBasicMaterial( { color: 0xffaacc, transparent: true, blending: THREE.AdditiveBlending } );
+      let material = new THREE.MeshPhongMaterial({ color: 0xdddddd, specular: 0xcccccc, shininess: 3, flatShading: true })
+
+      // this.mesh = new THREE.Mesh(geometry, material);
+      // //this.mesh.position.y = 0.2;
+      // this.scene.add(this.mesh);
+
+      // let texture = new THREE.Texture( this.generateTexture() );
+
+      for (var i = -8; i <= 8; i++) {
+        for (var j = -8; j <= 8; j++) {
+          // for(var k=-2; k<=2; k++) {
+          if (i === 1 && j === 1) {
+            let geometry = new THREE.SphereBufferGeometry(0.23, 32, 16)
+            let material = new THREE.MeshLambertMaterial({ map: texture, transparent: true })
+            let mesh = new THREE.Mesh(geometry, material)
+            mesh.position.x = i * 0.8
+            mesh.position.y = j * 0.8
+            this.scene.add(mesh)
+            this.mesh.push(mesh)
+          } else if (i === 0 && j === 1) {
+            // let texture = THREE.ImageUtils.loadTexture("src/images/ta.png",null,function(t) { });
+            let texture = new THREE.TextureLoader().load(ta)
+            let geometry = new THREE.SphereBufferGeometry(0.23, 32, 16)
+            let material = new THREE.MeshLambertMaterial({ map: texture, transparent: true })
+            let mesh = new THREE.Mesh(geometry, material)
+            mesh.position.x = i * 0.8
+            mesh.position.y = j * 0.8
+            this.scene.add(mesh)
+            this.mesh.push(mesh)
+          } else if (i === 1 && j === 0) {
+            let geometry = new THREE.SphereBufferGeometry(0.23, 15, 15)
+            let material = new THREE.MeshBasicMaterial({ color: 0xffaa00, wireframe: true })
+            let mesh = new THREE.Mesh(geometry, material)
+            mesh.position.x = i * 0.8
+            mesh.position.y = j * 0.8
+            this.scene.add(mesh)
+            this.mesh.push(mesh)
+          } else if (i === -1 && j === 1) {
+            let geometry = new THREE.SphereBufferGeometry(0.23, 35, 35)
+            let material = new THREE.MeshPhongMaterial({ color: 0x000000, specular: 0x666666, emissive: 0x333333, shininess: 10, opacity: 0.9, transparent: true })
+            let mesh = new THREE.Mesh(geometry, material)
+            mesh.position.x = i * 0.8
+            mesh.position.y = j * 0.8
+            this.scene.add(mesh)
+            this.mesh.push(mesh)
+          } else if (i === -1 && j === -1) {
+            // shadow
+            var canvas = document.createElement('canvas')
+            canvas.width = 2
+            canvas.height = 2
+            var context = canvas.getContext('2d')
+            var gradient = context.createRadialGradient(canvas.width / 2, canvas.height / 2, 0, canvas.width / 2, canvas.height / 2, canvas.width / 2)
+            gradient.addColorStop(0.1, 'rgba(210,210,210,1)')
+            gradient.addColorStop(1, 'rgba(255,255,255,1)')
+            context.fillStyle = gradient
+            context.fillRect(0, 0, canvas.width, canvas.height)
+            var shadowTexture = new THREE.CanvasTexture(canvas)
+
+            let geometry = new THREE.SphereBufferGeometry(0.23, 32, 16)
+            let material = new THREE.MeshLambertMaterial({ map: shadowTexture, transparent: true })
+            let mesh = new THREE.Mesh(geometry, material)
+            mesh.position.x = i * 0.8
+            mesh.position.y = j * 0.8
+            this.scene.add(mesh)
+            this.mesh.push(mesh)
+          } else {
+            let mesh = new THREE.Mesh(geometry, material)
+            mesh.position.x = i * 0.8
+            mesh.position.y = j * 0.8
+            mesh.position.z = Math.random()
+            // mesh.position.z = k*1;
+            this.scene.add(mesh)
+            if (i === 0 && j === 0) {
+              this.mesh.push(mesh)
+            }
+            this.mesh.push(mesh)
+          }
+          // }
+        }
+      }
+
+      // Lights
+      this.scene.add(new THREE.AmbientLight(0x222222))
+      var directionalLight = new THREE.DirectionalLight(0xaaaaaa, 0.125)
+      directionalLight.position.x = Math.random() - 0.5
+      directionalLight.position.y = Math.random() - 0.5
+      directionalLight.position.z = Math.random() - 0.5
+      directionalLight.position.normalize()
+      this.scene.add(directionalLight)
+      this.pointLight = new THREE.PointLight(0xaaaaaa, 1)
+      this.pointLight.position.z = 5
+      this.scene.add(this.pointLight)
+
+      window.addEventListener('resize', this.onWindowResize, false)
+    },
+    animate: function () {
+      var timer = 0.0001 * Date.now()
+      this.camera.position.x = Math.cos(timer) * 2
+      // this.camera.position.y = Math.sin( timer ) * 2;
+      this.camera.position.z = Math.sin(timer) * 2
+      this.camera.lookAt(this.scene.position)
+      for (var i = 0, l = this.mesh.length; i < l; i++) {
+        var mesh = this.mesh[ i ]
+        mesh.rotation.x += 0.01
+        mesh.rotation.y += 0.005
+      }
+
+      this.pointLight.position.x = Math.sin(timer * 7) * 300
+      this.pointLight.position.y = Math.cos(timer * 5) * 400
+      this.pointLight.position.z = Math.cos(timer * 3) * 300
+
+      requestAnimationFrame(this.animate)
+      this.controls.update()
+      this.renderer.render(this.scene, this.camera)
+    }
+  },
+  mounted () {
+    this.init()
+    this.animate()
   }
+}
 </script>
 <style rel="stylesheet/scss" lang="scss">
   @import "../../styles/mixin.scss";
@@ -316,13 +315,22 @@
       color: white;
     }
     .login-form {
-      position: absolute;
-      top: 0;
-      left: 0;
-      right: 0;
+      /*position: absolute;*/
+      /*top: 0;*/
+      /*left: 0;*/
+      /*right: 0;*/
+      /*width: 400px;*/
+      /*margin: auto;*/
+      /*padding: 35px 35px 15px 35px;*/
+      /*margin: 120px auto;*/
+
       width: 400px;
-      padding: 35px 35px 15px 35px;
-      margin: 120px auto;
+      margin: 0;
+      position: absolute;
+      top: 50%;
+      left: 50%;
+      -ms-transform: translate(-50%, -50%);
+      transform: translate(-50%, -50%);
       z-index: 1000000;
     }
     .el-form-item {
